// src/app/shared/interfaces/auxilium/work-order-center/work-order-center.interface.ts
export interface WorkOrderDisplay {
    id: number;
    patientid: number;
    itemid: number;
    assetid: number;
    patientpayorid1: number;
    patientpayorid2: number;
    salesid: number;
    referid: number;
    cmnformid: number;
    physician: string;
    patientdoctorid1: number;
    facilityid: number;
    sourceid: number;
    sourcetable: string;
    branchid: number;
    invoiceno: string;
    patalpha: string;
    lastweekbilled: number;
    lastdatebilled: string;
    previousdatebilled: string;
    svcdate: string;
    todate: string;
    recordtype: string;
    paponumber: string;
    diag24e: string;
    diagnosis: string;
    patienticdid1: number;
    patienticdid2: number;
    patienticdid3: number;
    patienticdid4: number;
    submitted: number;
    allowed: number;
    dueprimary: number;
    duesecondary: number;
    prepaid: number;
    prepaid2: number;
    pos: string;
    tos: string;
    billtype: string;
    stopcode: string;
    priceid: number;
    quantity: number;
    stopdate: string;
    authdate: string;
    cmnexpire: string;
    nextbill: string;
    hcpcscode: string;
    mod1: string;
    mod2: string;
    mod3: string;
    mod4: string;
    physeq: string;
    cmnreq: string;
    firstcmn: string;
    nextcmn: string;
    cmnfreq: number;
    description: string;
    monthsrented: number;
    entrydate: string;
    opid: string;
    assignment: boolean;
    ytdbilled: number;
    pickupreason: string;
    narrative: string;
    outsidelab: string;
    charges: number;
    shipstatus: string;
    shipdate: string;
    billcycle: string;
    admit: string;
    discharge: string;
    confirm: string;
    confirmid: string;
    lotnumber: string;
    orderedby: string;
    ordphone: string;
    deliverytype: string;
    dateprinted: string;
    groupid: string;
    otherhold: string;
    calories: number;
    tenmonth: string;
    checkno: string;
    diagnos2: string;
    diagnos3: string;
    diagnos4: string;
    adduserid: string;
    adddate: string;
    followup: string;
    primbill: string;
    secbill: string;
    isoverride: boolean;
    chcup: string;
    lineno: number;
    paymenttype: string;
    paymenttype2: string;
    authcount: number;
    authused: number;
    invtcat2update: string;
    notes: string;
    inventorylocationid: number;
    inventorybinid: number;
    fld2876: string;
    patienticdid5: number;
    patienticdid6: number;
    patienticdid7: number;
    patienticdid8: number;
    description2: string;
    trackingno: string;
    shipvia: string;
    quantityordered: number;
    quantityshipped: number;
    usagelimitquantity: number;
    lastuser: string;
    datedelivered: string;
    lot: boolean;
    upnqualifier: string;
    payorid1: number;
    payorid2: number;
    doctorid1: number;
}

export type GetWorkOrderRepsResponse = WorkOrderDisplay[];